Encode item components on entity remove rather than periodically, clear cache on quit event
