+++ geary-papermc-tracking/src/main/kotlin/com/mineinabyss/geary/papermc/tracking/items/ItemTracking.kt
@@ -27,7 +27,7 @@ interface ItemTracking {
             gearyPaper.plugin.listeners(LoginListener())
             geary.pipeline.addSystems(
                 InventoryTrackerSystem(),
-                PeriodicSaveSystem(),
+//                PeriodicSaveSystem(),
                 CustomModelDataToPrefabTracker()
             )
         }
